coordinator/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
supabase==2.11.0
pydantic==2.10.4
python-dotenv==1.0.1
